added close open button for namespace and added comments
