+++ src/app/ui/PodModal.tsx
@@ -1,13 +1,18 @@
 import React from 'react';
 
-
+//Modal component that displays when moreinfo is clicked on podcard 
 export default function Modal({ isOpen, onClose, children }) {
+  //checks if isOpen prop is false, nothing will be rendered if isOpen false 
   if (!isOpen) return null;
 
+  //when isOpen is true
   return (
+    //modal box with the white background 
     <div className='fixed inset-0 bg-black bg-opacity-50 flex justify-center items-center z-50'>
+      {/* section that renders children props  */}
       <div className='bg-white p-6 rounded-lg shadow-xl'>
         {children}
+        {/* button for closing the modal. When clicked it invokes the 'onClose' function */}
         <button
           onClick={onClose}
           className='mt-4 bg-red-500 hover:bg-red-700 text-white py-1 px-2 rounded'
