+++ types/types.ts
@@ -1,5 +1,18 @@
 type PieChartData = { [key: string]: number };
 
+export interface PodItem {
+  status: string;
+  namespace: string;
+  name: string;
+  creationTimestamp: Date;
+  dnsPolicy: string;
+  containers: number;
+  restartPolicy: string;
+  hostIP: string;
+  podIP: string;
+  startTime: Date;
+}
+
 type PromMetricsData = any;
 type KubeMetricsstatus = any;
 type KubePodsStatus = any;
